+++ Src/Library/Extensions/StreamHelper.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Http.Headers;
 using Microsoft.Extensions.Primitives;
 using Microsoft.Net.Http.Headers;
+using System.Net.Mail;
 
 namespace FastEndpoints;
 
@@ -36,7 +37,11 @@ internal static class StreamHelper
         response.ContentType = contentType;
 
         if (fileName is not null)
-            ctx.Response.Headers.Add("Content-Disposition", $"attachment; filename={fileName}");
+        {
+            var cdHdr = new ContentDispositionHeaderValue("attachment");
+            cdHdr.SetHttpFileName(fileName);
+            ctx.Response.Headers.ContentDisposition = cdHdr.ToString();
+        }
 
         if (fileLength is not null)
         {